+++ backend/core/src/main/java/com/rollthedice/backend/domain/debate/api/DebateController.java
@@ -68,6 +68,6 @@ public class DebateController implements DebateApi {
     @GetMapping("/summary/{roomId}")
     @Override
     public DebateSummaryResponse getSummarizedDebate(@PathVariable final Long roomId) {
-        return debateMessageService.summaryDebateMessages(roomId);
+        return debateRoomService.summaryDebateMessages(roomId);
     }
 }
